backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0

# Math / Quant
numpy==1.26.4
scipy==1.13.0
pandas==2.2.2

# HTTP / Data
httpx==0.27.0
aiohttp==3.9.5
requests==2.32.2
beautifulsoup4==4.12.3
lxml==5.2.1

# Cache
redis==5.0.4
hiredis==2.3.2

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Payments
razorpay==1.4.1

# WebSocket
websockets==12.0

# Logging
structlog==24.2.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0

# Dev
black==24.4.2
isort==5.13.2
